+++ services/token.go
@@ -17,7 +17,7 @@ func NewTokenService(db *pgxpool.Pool) *TokenService {
 	}
 }
 
-func (s *TokenService) GetRefreshToken(ctx context.Context, userId int) (string, error) {
+func (s *TokenService) GetRefreshToken(ctx context.Context, userId int64) (string, error) {
 	var token string
 	err := s.db.QueryRow(ctx, "SELECT token FROM refresh_tokens WHERE user_id = $1", userId).Scan(&token)
 	if err != nil {
@@ -27,7 +27,7 @@ func (s *TokenService) GetRefreshToken(ctx context.Context, userId int) (string,
 	return token, nil
 }
 
-func (s *TokenService) AddRefreshToken(ctx context.Context, userId int, token string, exp time.Time) error {
+func (s *TokenService) AddRefreshToken(ctx context.Context, userId int64, token string, exp time.Time) error {
 	_, err := s.db.Exec(ctx, "INSERT INTO refresh_tokens (user_id, token, expires_at) VALUES ($1, $2, $3)", userId, token, exp)
 	if err != nil {
 		return err
@@ -35,7 +35,7 @@ func (s *TokenService) AddRefreshToken(ctx context.Context, userId int, token st
 	return nil
 }
 
-func (s *TokenService) RemoveRefreshToken(ctx context.Context, userId int) error {
+func (s *TokenService) RemoveRefreshToken(ctx context.Context, userId int64) error {
 	_, err := s.db.Exec(ctx, "DELETE FROM refresh_tokens WHERE user_id = $1", userId)
 	if err != nil {
 		return err
